+++ CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - Add root admin route and check if page requires auth before navigate
 - Add admin component SignIn
 - Add admin component SignUp
+- Add demo admin component AdminHeader
 ### Changed
 - Updated page store with action returning a promise
 ### Deprecated